quiz/requirement.txt
django
djangorestframework
django-nested-inline